refactor all reviews dto and mappers to delete duplicated
