+++ repository/media/src/main/java/com/giraffe/media/movie/datasource/remote/MoviesRemoteDataSource.kt
@@ -1,9 +1,9 @@
 package com.giraffe.media.movie.datasource.remote
 
-import  com.giraffe.media.movie.datasource.remote.dto.MovieDto
-import  com.giraffe.media.movie.datasource.remote.dto.MovieGenreDto
-import  com.giraffe.media.movie.datasource.remote.dto.MovieReviewDto
-import  com.giraffe.media.movie.datasource.remote.dto.RatingRequest
+import com.giraffe.media.dto.ReviewDto
+import com.giraffe.media.movie.datasource.remote.dto.MovieDto
+import com.giraffe.media.movie.datasource.remote.dto.MovieGenreDto
+import com.giraffe.media.movie.datasource.remote.dto.RatingRequest
 
 interface MoviesRemoteDataSource {
 
@@ -17,7 +17,7 @@ interface MoviesRemoteDataSource {
 
     suspend fun getMoviesByGenre(genreId: Int, page: Int): List<MovieDto>
 
-    suspend fun getMovieReviews(movieId: Int): List<MovieReviewDto>
+    suspend fun getMovieReviews(movieId: Int): List<ReviewDto>
 
     suspend fun getMovieRecommendations(movieId: Int, page: Int): List<MovieDto>
